refactor(skin): split preview responsibilities
